CensusApp/backend/requirements.txt
PyMySQL==1.1.1
blinker==1.9.0
click==8.1.7
Flask==3.1.0
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
watchdog==6.0.0
Werkzeug==3.1.3
requests==2.31.0
python-dotenv==1.0.0
flask-caching==2.1.0
cryptography==41.0.1
pandas==2.1.0
PyMySQL==1.1.0